+++ src/plugins/sliceGrouper.ts
@@ -1,0 +1,1 @@
+export { SliceGrouper } from "../.internal/plugins/sliceGrouper/SliceGrouper";